switch

+++ api/compiler/ValueWrapper.cs
@@ -4,7 +4,7 @@ public abstract record ValueWrapper;
 
 public record IntValue (int Value) : ValueWrapper;
 
-public record FloatValue (float Value) : ValueWrapper;
+public record FloatValue (double Value) : ValueWrapper;
 public record StringValue (string Value) : ValueWrapper;
 
 public record BoolValue (bool Value) : ValueWrapper;
